perf: restructure react components
