+++ src/main/pages/src/components/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Login from './Login';
 import SignUp from './SignUp';
+import ToolBar from '../shared/ToolbarView'
 import { HashRouter, Route } from 'react-router-dom';
 
 class App extends React.Component {
@@ -10,6 +11,7 @@ class App extends React.Component {
                 <div>
                     <Route exact path='/' component={Login} />
                     <Route path="/SignUp" component={SignUp} />
+                    <Route path="/ToolBar" component={ToolBar}/>
                 </div>
             </HashRouter>
         );
